requirements.txt
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
pandas==2.0.1
pyspm==0.3.0
python-dotenv==1.0.0
requests==2.31.0